refactor: centralize server into ServerContext
This commit centralizes all server related logic into a central
 ServerContext class.
